Dados do censo ensino superior INEP 2017 (Cursos)

Base de dados do INEP para o censo brasileiro do ensino superior (cursos).

References

http://inep.gov.br/microdados

Author

INEP

Examples

require(rnp)
str(dm_curso)
#> Classes ‘data.table’ and 'data.frame':	35693 obs. of  53 variables:
#>  $ NU_ANO_CENSO                    : int  2017 2017 2017 2017 2017 2017 2017 2017 2017 2017 ...
#>  $ CO_IES                          : int  789 4567 2341 670 403 1058 457 275 1563 3610 ...
#>  $ CO_LOCAL_OFERTA                 : int  1033528 659871 131728 1786 657876 149637 1002303 2000174 658582 1065378 ...
#>  $ CO_UF                           : int  14 51 35 35 53 29 35 35 22 53 ...
#>  $ CO_MUNICIPIO                    : int  1400100 5107925 3510609 3509502 5300108 2927408 3547809 3549904 2211001 5300108 ...
#>  $ CO_CURSO                        : int  1259131 1258115 68433 52052 18061 1288733 1303755 6578 38315 114977 ...
#>  $ NO_CURSO                        : chr  "MÚSICA" "GESTÃO DE RECURSOS HUMANOS" "DIREITO" "ENGENHARIA ELÉTRICA" ...
#>  $ CO_OCDE_AREA_GERAL              : int  1 3 3 5 3 7 4 1 7 3 ...
#>  $ CO_OCDE_AREA_ESPECIFICA         : int  14 34 38 52 38 72 42 14 72 34 ...
#>  $ CO_OCDE_AREA_DETALHADA          : int  146 345 380 522 380 723 421 145 723 342 ...
#>  $ CO_OCDE                         : chr  "146F20" "345G09" "380D01" "522E06" ...
#>  $ NU_CARGA_HORARIA                : int  3210 1720 4080 3892 3790 4360 3200 2849 4000 1880 ...
#>  $ DT_INICIO_FUNCIONAMENTO         : chr  "23/09/2013" "25/01/2016" "16/02/2004" "01/08/2001" ...
#>  $ NU_INTEGRALIZACAO_NOTURNO       : num  4 2.5 5 5 5 5 4 3 NA 2 ...
#>  $ QT_INSC_VAGA_NOVA_NOTURNO       : int  469 52 50 110 1219 481 557 23 NA 41 ...
#>  $ QT_INSC_VAGA_REMAN_NOTURNO      : int  0 28 30 15 99 48 4 3 NA 0 ...
#>  $ QT_INSC_PROG_ESP_NOTURNO        : int  0 0 0 0 0 0 0 0 NA 0 ...
#>  $ QT_VAGAS_NOVAS_NOTURNO          : int  50 50 100 67 360 100 70 40 NA 100 ...
#>  $ QT_VAGAS_REMAN_NOTURNO          : int  0 45 20 124 60 472 1 10 NA 0 ...
#>  $ QT_VAGAS_PROG_ESP_NOTURNO       : int  0 0 0 0 0 0 0 0 NA 0 ...
#>  $ QT_MATRICULA_TOTAL              : int  131 12 247 172 1966 250 108 31 626 0 ...
#>  $ QT_CONCLUINTE_TOTAL             : int  5 0 71 26 450 0 0 4 191 0 ...
#>  $ QT_INGRESSO_TOTAL               : int  40 4 51 35 223 116 147 8 244 0 ...
#>  $ QT_INGRESSO_VAGA_NOVA           : int  40 3 46 33 179 77 146 5 139 0 ...
#>  $ QT_VAGA_TOTAL                   : int  50 95 240 191 840 1143 150 90 337 100 ...
#>  $ DESC_IN_AJUDA_DEFICIENTE        : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_CAPITAL                 : chr  "1. Sim" "0. Não" "0. Não" "0. Não" ...
#>  $ DESC_IN_DISCIPLINA_LIBRAS       : chr  "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" ...
#>  $ DESC_IN_GRATUITO                : chr  "1. Sim" "0. Não" "0. Não" "0. Não" ...
#>  $ DESC_IN_GUIA_INTERPRETE         : chr  "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" ...
#>  $ DESC_IN_INTEGRAL                : chr  "0. Não" "0. Não" "0. Não" "0. Não" ...
#>  $ DESC_IN_MATERIAL_AMPLIADO       : chr  "1. Sim" "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" ...
#>  $ DESC_IN_MATERIAL_AUDIO          : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_MATERIAL_BRAILLE        : chr  "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" ...
#>  $ DESC_IN_MATERIAL_DIGITAL        : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_MATERIAL_IMPRESSO       : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_MATERIAL_LIBRAS         : chr  "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" ...
#>  $ DESC_IN_MATERIAL_TATIL          : chr  "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" ...
#>  $ DESC_IN_MATUTINO                : chr  "0. Não" "0. Não" "1. Sim" "0. Não" ...
#>  $ DESC_IN_NOTURNO                 : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_OFERECE_DISC_SEMI_PRES  : chr  "0. Não" "0. Não" "0. Não" "0. Não" ...
#>  $ DESC_IN_POSSUI_LABORATORIO      : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_RECURSOS_COMUNICACAO    : chr  "1. Sim" "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" ...
#>  $ DESC_IN_RECURSOS_INFORMATICA    : chr  "1. Sim" "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" ...
#>  $ DESC_IN_TRADUTOR_LIBRAS         : chr  "1. Sim" "1. Sim" "0. Não (Curso possui outro tipo de ajuda ao aluno com deficiência)" "1. Sim" ...
#>  $ DESC_IN_VESPERTINO              : chr  "0. Não" "0. Não" "0. Não" "0. Não" ...
#>  $ DESC_TP_ATRIBUTO_INGRESSO       : chr  "0. Normal" "0. Normal" "0. Normal" "0. Normal" ...
#>  $ DESC_TP_CATEGORIA_ADMINISTRATIVA: chr  "1. Pública Federal" "5. Privada sem fins lucrativos" "4.Privada com fins lucrativos" "5. Privada sem fins lucrativos" ...
#>  $ DESC_TP_GRAU_ACADEMICO          : chr  "2. Licenciatura" "3. Tecnológico" "1. Bacharelado" "1. Bacharelado" ...
#>  $ DESC_TP_MODALIDADE_ENSINO       : chr  "1. Presencial" "1. Presencial" "1. Presencial" "1. Presencial" ...
#>  $ DESC_TP_NIVEL_ACADEMICO         : chr  "1. Graduação" "1. Graduação" "1. Graduação" "1. Graduação" ...
#>  $ DESC_TP_ORGANIZACAO_ACADEMICA   : chr  "1. Universidade" "3. Faculdade" "3. Faculdade" "1. Universidade" ...
#>  $ DESC_TP_SITUACAO                : chr  "1. Em atividade" "1. Em atividade" "1. Em atividade" "1. Em atividade" ...
#>  - attr(*, ".internal.selfref")=<pointer: (nil)>