Imprime um resultado do rnp

Metodo de impressao para os objetos de classe rnp_resultado (listas de componentes devolvidas por varias funcoes do pacote). Imprime cada componente de forma legivel: tabelas sao mostradas; objetos-modelo e graficos sao resumidos em uma linha; escalares aparecem em linha.

Usage

# S3 method for class 'rnp_resultado'
print(x, ...)

Arguments

x

Objeto de classe rnp_resultado.

...

Ignorado.

Value

x, de forma invisivel.