rnp: Recursos estatisticos do projeto R NA PRATICA

Reune as funcoes estatisticas do projeto R NA PRATICA, organizadas por area: analise descritiva, probabilidade, inferencia, regressao, multivariada, series temporais, amostragem e dados categoricos. As funcoes analiticas retornam saidas em formato tibble e os graficos usam ggplot2. Rotinas numericamente intensivas sao escritas em C++ (Rcpp/RcppArmadillo). Inclui tambem ferramentas para os microdados do censo do ensino superior do INEP.

Modulos

rnp_descritiva(), rnp_quantis(), rnp_outliers()

rnp_tabela_classes(), rnp_distribuicao()

rnp_ic_media(), rnp_teste_t(), rnp_anova()

rnp_correlacao_teste(), rnp_regressao()

rnp_pca(), rnp_kmeans()

rnp_ts_decomposicao(), rnp_ts_acf(), rnp_media_movel()

rnp_combinacao(), rnp_distribuicao_binomial()

rnp_grafico_dispersao(), rnp_tema_rnp()

rnp_amostra_simples(), rnp_tamanho_amostra_media()

rnp_tamanho_efeito(), rnp_na_summary()

See also

Author

Jose E. Lopes evandeilton@gmail.com