Teorema de Bayes (particao do espaco amostral)

Calcula as probabilidades a posteriori P(H_i | E) a partir das probabilidades a priori P(H_i) e das verossimilhancas P(E | H_i), para uma particao de hipoteses mutuamente exclusivas e exaustivas. O caso classico de dois eventos e obtido com vetores de comprimento 2.

Usage

rnp_bayes(priori, verossimilhanca, hipoteses = NULL, digits = 4L)

Arguments

priori

Vetor de probabilidades a priori (deve somar 1).

verossimilhanca

Vetor P(E | H_i), mesmo comprimento de priori.

hipoteses

Vetor opcional de rotulos das hipoteses.

digits

Inteiro. Casas decimais.

Value

tibble com hipotese, priori, verossimilhanca, conjunta (P(H_i e E)) e posteriori (P(H_i | E)).

See also

Examples

# Teste diagnostico: prevalencia 1%, sensibilidade 99%, especificidade 95%
rnp_bayes(priori = c(doente = 0.01, sadio = 0.99),
          verossimilhanca = c(0.99, 0.05))
#> # A tibble: 2 × 5
#>   hipotese priori verossimilhanca conjunta posteriori
#>   <chr>     <dbl>           <dbl>    <dbl>      <dbl>
#> 1 doente     0.01            0.99   0.0099      0.167
#> 2 sadio      0.99            0.05   0.0495      0.833