Cadeia de Markov de tempo discreto

A partir de uma matriz de transicao, calcula a distribuicao apos n passos, a matriz de transicao em n passos (P^n) e a distribuicao estacionaria. Usa backend C++ (RcppArmadillo).

Usage

rnp_cadeia_markov(P, estado_inicial = 1L, n = 10L, digits = 4L)

Arguments

P

Matriz de transicao (linhas somam 1).

estado_inicial

Vetor de distribuicao inicial (soma 1) ou indice inteiro do estado de partida.

n

Inteiro. Numero de passos.

digits

Inteiro. Casas decimais.

Value

Uma lista com:

distribuicao_n: tibble com a distribuicao apos n passos.

estacionaria: tibble com a distribuicao estacionaria.

matriz_n: matriz P^n.

See also

Examples

P <- matrix(c(0.9, 0.1, 0.5, 0.5), 2, 2, byrow = TRUE)
rnp_cadeia_markov(P, estado_inicial = 1, n = 10)
#> $distribuicao_n
#> # A tibble: 2 × 2
#>   estado probabilidade
#>   <chr>          <dbl>
#> 1 E1             0.833
#> 2 E2             0.167
#> 
#> $estacionaria
#> # A tibble: 2 × 2
#>   estado probabilidade
#>   <chr>          <dbl>
#> 1 E1             0.833
#> 2 E2             0.167
#> 
#> $matriz_n
#>        [,1]   [,2]
#> [1,] 0.8334 0.1666
#> [2,] 0.8332 0.1668
#>