Analise de correspondencia (CA)

Decompoe uma tabela de contingencia para mapear linhas e colunas em um espaco de baixa dimensao (via SVD dos residuos padronizados).

Usage

rnp_correspondencia(tabela, n_dim = 2L, digits = 4L)

Arguments

tabela

Matriz/tabela de contingencia (contagens).

n_dim

Inteiro. Numero de dimensoes a reter.

digits

Inteiro.

Value

Uma lista com inercia (tibble), coord_linhas (tibble) e coord_colunas (tibble).

See also

Examples

rnp_correspondencia(table(mtcars$cyl, mtcars$gear))
#> $inercia
#> # A tibble: 2 × 3
#>   dimensao inercia   prop
#>      <int>   <dbl>  <dbl>
#> 1        1  0.563  0.998 
#> 2        2  0.0011 0.0019
#> 
#> $coord_linhas
#> # A tibble: 3 × 3
#>   categoria   Dim1    Dim2
#>   <chr>      <dbl>   <dbl>
#> 1 4         -0.799  0.0286
#> 2 6         -0.415 -0.0587
#> 3 8          0.835  0.0069
#> 
#> $coord_colunas
#> # A tibble: 3 × 3
#>   categoria    Dim1    Dim2
#>   <chr>       <dbl>   <dbl>
#> 1 3          0.746  -0.0123
#> 2 4         -0.895  -0.016 
#> 3 5         -0.0913  0.0754
#>