Teste de Dickey-Fuller aumentado (ADF)

Testa a presenca de raiz unitaria (H0: serie nao-estacionaria). Ajusta a regressao da serie diferenciada sobre o nivel defasado e sobre defasagens da propria diferenca (forma "aumentada"), e compara a estatistica do termo de nivel com valores criticos de Dickey-Fuller.

Usage

rnp_ts_adf(x, lag = NULL, tipo = c("constante", "tendencia"), digits = 4L)

Arguments

x

Vetor numerico.

lag

Numero de defasagens aumentadas. Default automatico.

tipo

String: "constante" (default) ou "tendencia".

digits

Inteiro.

Value

tibble com estatistica, lag, valor_critico_5, p_valor_aprox, estacionaria.

See also

Examples

rnp_ts_adf(as.numeric(lh))
#> # A tibble: 1 × 5
#>   estatistica   lag valor_critico_5 p_valor_aprox estacionaria
#>         <dbl> <dbl>           <dbl>         <dbl> <lgl>       
#> 1       -2.96     3           -2.86         0.043 TRUE